requirements.txt
adafruit-circuitpython-rgb-display
Pillow
soco
evdev
